Declare exchanges/queues independent from client (fix #2152)
